requirements.txt
bottle==0.12.8
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
Pillow==8.1.2
playsound==1.2.2
pylab-sdk==1.3.2
pyparsing==2.4.7
pyrtlsdr==0.2.92
python-dateutil==2.8.1
scipy==1.6.1
six==1.15.0
